Add files via upload
Nabila is working for the Department for Transportation in the United Kingdom.
One day, a huge festival is about to take place in Hogwarts, Scotland. To handle the
traffic, Nabila wants to use a special traffic light in an intersection near the festival
grounds that cycles in order:
Green -> Yellow -> Red -> (back to Green)
Due to one thing and another, the light does not start in green, instead it starts in
another color and the cycle continues normally after that.
The duration of green light is 20 seconds, yellow light 3 seconds, and red light 80
seconds. If in the 45th second the light just showed yellow, determine a logic that
could retrieve the color of the traffic light in:
80th second
135th second
150th second
212nd second
After determining the logic, find the color shown by the traffic
light on the given time using your logic!
Clue: The light shows yellow on the 45th to the 47th second
(This source code file contains the solution for the description above)
(hijau = green, kuning = yellow, red = merah)